requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
etils==1.13.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.2.0
glfw==2.9.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
grpcio==1.65.5
h5py==3.11.0
icecream==2.1.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.3
ipython==8.18.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.0.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.23.1
oauthlib==3.3.1
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
path==17.1.1
pillow==11.0.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
protobuf==3.19.6
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.1
Pygments==2.18.0
pyhocon==0.3.57
PyMCubes==0.1.6
PyOpenGL==3.1.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==9.1.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2024.8.10
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tqdm==4.66.5
traitlets==5.14.3
trimesh==4.4.7
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.11
zipp==3.20.0
